Add script to generate js from svg
This adds a script to generate the js frontend file from the svg files.
This new script supports the new icon picker.
The script uses the filebasename (e.g. icon from icon.svg) as icon name.
To add keywords for the icon picker search add a desc tag to your svg
and fill it with space separated keywords.
e.g. <desc>light ceiling</desc>
The script will only pick up the first desc tag of the document.
If the tag is missing the keywords will be set to an empty list.
The script will also only pick up the first path as the icons path so
make sure they are all single path icons.